+++ app-android/src/main/java/com/wolfmontwe/mhp/challenge/mobile/android/app/presentation/feature/house/list/HouseCardItem.kt
@@ -63,8 +63,15 @@ fun HouseCardItemPreview() {
     MainTheme {
         HouseCardItem(
             House(
-                id = Identifier("378"),
-                name = "House Targaryen of King's Landing",
+                id = Identifier("123"),
+                name = "TestName",
+                region = "TestRegion",
+                coatOfArms = "TestCoatOfArms",
+                words = "TestWords",
+                titles = listOf("TestTitle1", "TestTitle2"),
+                seats = listOf("TestSeat1", "TestSeat2"),
+                members = 11,
+                founded = "TestFounded"
             ),
             {}
         )