+++ app-android/src/main/java/com/wolfmontwe/mhp/challenge/mobile/android/app/presentation/feature/house/list/HouseListContent.kt
@@ -42,8 +42,8 @@ fun HouseListContent(
             }
             HouseCardItem(house = item, onClick = { onClickHouse(item.id.value) })
         }
-        if (isLoading) {
-            item {
+        item {
+            if (isLoading) {
                 Row(
                     modifier = Modifier
                         .fillMaxWidth()
@@ -74,8 +74,15 @@ fun HouseListContentPreview() {
             items = listOf(
                 House(
                     id = Identifier("123"),
-                    name = "House Test"
-                )
+                    name = "TestName",
+                    region = "TestRegion",
+                    coatOfArms = "TestCoatOfArms",
+                    words = "TestWords",
+                    titles = listOf("TestTitle1", "TestTitle2"),
+                    seats = listOf("TestSeat1", "TestSeat2"),
+                    members = 11,
+                    founded = "TestFounded"
+                ),
             ),
             isLoading = true,
             onClickHouse = {},